# Improve perf of Rosetta get genesis balance script