--- a/hedera-mirror-rosetta/scripts/validation/get-genesis-balance.sh
+++ b/hedera-mirror-rosetta/scripts/validation/get-genesis-balance.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 set -euo pipefail
 
+# handle input argument defaults
+network=${1:-demo}
+account_limit=${2:-20}
+starting_timestamp=${3:-0}
+
 currency=$(cat <<EOF
 {
   "symbol": "HBAR",
@@ -11,14 +16,22 @@ currency=$(cat <<EOF
 }
 EOF
 )
-genesis_timestamp_query="select consensus_timestamp from account_balance_file order by consensus_timestamp asc limit 2"
 
-genesis_hbar_balance_query=$(cat <<EOF
-\set ON_ERROR_STOP on
+genesis_timestamp_query=$(cat <<EOF
+select consensus_timestamp
+from account_balance_file
+where consensus_timestamp > :starting_timestamp
+order by consensus_timestamp asc
+limit 1
+EOF
+)
+
+one_week_ns=604800000000000
+applicable_accounts_query=$(cat <<EOF
 with recent_crypto_accounts as (
  select distinct(entity_id)
- from crypto_transfer where consensus_timestamp > :genesis_timestamp and consensus_timestamp <= :second_timestamp
- limit 20
+ from crypto_transfer where consensus_timestamp > :genesis_timestamp and consensus_timestamp <= :genesis_timestamp + :one_week_ns
+ limit :account_limit
 ),
 genesis_balance as (
   select account_id, balance
@@ -28,12 +41,47 @@ genesis_balance as (
   where balance <> 0 and ab.consensus_timestamp = :genesis_timestamp
   group by account_id,balance
 )
+EOF
+)
+
+if [[ "$account_limit" == "0" ]]; then
+  echo "Account limit removed, all accounts with non-zero balance account in initial balance file will be considered"
+  applicable_accounts_query=$(cat <<EOF
+    with genesis_balance as (
+      select account_id, balance
+      from account_balance
+      where balance <> 0 and consensus_timestamp = :genesis_timestamp
+      group by account_id, balance
+    )
+EOF
+  )
+fi
+
+genesis_hbar_balance_query=$(cat <<EOF
+\set ON_ERROR_STOP on
+${applicable_accounts_query}
 select json_agg(json_build_object('id', account_id::text, 'value', balance::text))
 from genesis_balance
 EOF
 )
 
-network=${1:-demo}
+genesis_token_balance_query=$(cat <<EOF
+\set ON_ERROR_STOP on
+select json_agg(json_build_object(
+  'id', account_id::text,
+  'token', tb.token_id::text,
+  'decimals', t.decimals,
+  'value', balance::text,
+  'type', t.type
+))
+from token_balance tb
+join token t on t.token_id = tb.token_id
+where tb.consensus_timestamp = :genesis_timestamp and
+  tb.account_id in (:account_ids) and
+  balance <> 0
+EOF
+)
+
 parent_path="$(cd "$(dirname "${BASH_SOURCE[0]}")" ; pwd -P)"
 psql_cmd="psql -h localhost -U mirror_rosetta mirror_node -t -P format=unaligned"
 
@@ -42,12 +90,16 @@ echo "localhost:5432:mirror_node:mirror_rosetta:mirror_rosetta_pass" > ~/.pgpass
 SECONDS=0
 while [[ $SECONDS -lt 120 ]];
 do
-  first_two_timestamps=($($psql_cmd -c "$genesis_timestamp_query"))
-  genesis_timestamp=${first_two_timestamps[0]}
-  second_timestamp=${first_two_timestamps[1]}
+  genesis_timestamp=$(echo "$genesis_timestamp_query" | $psql_cmd -v starting_timestamp="$starting_timestamp")
+  echo "Retrieved genesis balance file timestamp ${genesis_timestamp}"
+  if [[ -z "$genesis_timestamp" ]]; then
+    echo "Failed to get genesis timestamp"
+    exit 1
+  fi
+
   if [[ -n "$genesis_timestamp" ]]; then
     # get genesis hbar balances from genesis account balance file
-    account_balances=$(echo "$genesis_hbar_balance_query" | $psql_cmd -v genesis_timestamp="$genesis_timestamp" -v second_timestamp="$second_timestamp")
+    account_balances=$(echo "$genesis_hbar_balance_query" | $psql_cmd -v genesis_timestamp="$genesis_timestamp" -v one_week_ns="$one_week_ns" -v account_limit="$account_limit")
     echo "account balances - $(echo "$account_balances" | jq . )"
     break
   fi
@@ -55,11 +107,6 @@ do
   echo "${SECONDS}s elapsed, retry getting genesis timestamp in 5 seconds"
   sleep 5
 done
-
-if [[ -z "$genesis_timestamp" ]]; then
-  echo "Failed to get genesis timestamp"
-  exit 1
-fi
 
 hbar_json=$(echo "$account_balances" | \
   jq --argjson currency "$currency" \