**Description**:
The `hedera-mirror-rosetta/scripts/validation/get-genesis-balance.sh` script performs heavy lifting logic to get a genesis script. This sometimes runs slow in large environments and needs streamlining.

- Update `genesis_timestamp_query` to 
  - use order and limit instead of `min()`
  - Add `recent_crypto_accounts` CTE to pull the recent account used in a crypto transfer since balance file
  - Update `genesis_balance` to use result of `recent_crypto_accounts`
  - Add timestamp gated logic to query to speed up
  - Add support for `account_limit` and  `transfer_window_ns` params
- Update README 
  - with explanation of additional params support
  - split out sections for easier reading

**Related issue(s)**:

Fixes #

**Notes for reviewer**:
Before: 121 s
After: 24 s

Table size 


table_name | row_estimate | index | toast | table | total
-- | -- | -- | -- | -- | --
crypto_transfer | 2.3496112e+07 | 1767 MB | NULL | 2510 MB | 4277 MB
account_balance | 2.624885e+07 | 1926 MB | 8192 bytes | 1412 MB | 3338 MB
transaction | 6.438296e+06 | 691 MB | 8192 bytes | 1571 MB | 2262 MB
token_balance | 8.535088e+06 | 356 MB | NULL | 529 MB | 886 MB
token_transfer | 1.620688e+06 | 199 MB | NULL | 180 MB | 379 MB
account_balance_file | 362 | 64 kB | 8192 bytes | 112 kB | 184 kB
nft | 9 | 32 kB | 8192 bytes | 8192 bytes | 48 kB
token | 0 | 16 kB | 8192 bytes | 8192 bytes | 32 kB

---
Run against mainnet data
Before: 14197s
After: 72s

**Checklist**

- [ ] Documented (Code comments, README, etc.)
- [x] Tested (unit, integration, etc.)