@steven-sheehy 2022-02-04
Has this been tested against a full mainnet dump (or on failover)? Need to know the full runtime and optimize it if it's still too slow.